Fix issues with casting of return type

+++ stdlib/cmlc/lib/Math/cos.java
@@ -4,6 +4,6 @@ import cmlc.Function;
 
 public class cos extends Function {
 	public Object apply(Object input) {
-		return Math.cos((Float) input);
+		return (float) Math.cos((Float) input);
 	}
 }
