+++ stdlib/cmlc/lib/Math/sin.java
@@ -4,6 +4,6 @@ import cmlc.Function;
 
 public class sin extends Function {
 	public Object apply(Object input) {
-		return Math.sin((Float) input);
+		return (float) Math.sin((Float) input);
 	}
 }
