+++ stdlib/cmlc/lib/Math/sqrt.java
@@ -4,6 +4,6 @@ import cmlc.Function;
 
 public class sqrt extends Function {
 	public Object apply(Object input) {
-		return Math.sqrt((Float) input);
+		return (float) Math.sqrt((Float) input);
 	}
 }
